Use strdup instead of strlen/strcpy dance.
On aarch64, we are getting a new warning from rttest saying
that the strcpy wasn't properly 0-terminating the string.
It is actually a false warning (since we are terminating it
on the line before), but it turns out that we can simplify
this whole thing by using strdup.  Do that here.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>