Fix up nonsensical handling of NULL in rttest_get_{params,statistics} (#107)
clang static analysis pointed out that when rttest_get_params
or rttest_get_statistics was called with a NULL pointer, there
was a dead store.  That's completely correct; with a NULL
pointer, it would overwrite the pointer, but of course what
got returned was nothing.  Mildly refactor these APIs to instead
return an error with a NULL pointer, and only return data
when there is a valid pointer passed.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>